+++ .env
@@ -1,0 +1,2 @@
+export PYTHONPATH="$PYTHONPATH:$PWD"
+
