+++ src/message_schemas/join_game.json
