+++ src/message_schemas/new_connection_prompt.json
