+++ src/message_schemas/start_game.json
